web: Move config to public/config
