+++ web/packages/core/src/internal/builder.ts
@@ -1,5 +1,9 @@
 import type { RuffleInstanceBuilder } from "../../dist/ruffle_web";
-import { BaseLoadOptions, Duration, SecsDuration } from "../load-options";
+import {
+    BaseLoadOptions,
+    Duration,
+    SecsDuration,
+} from "../public/config/load-options";
 
 /**
  * Checks if the given value is explicitly `T` (not null, not undefined)
